requirements.txt
pandas==1.1.2
numpy==1.19.2
scikit-learn==0.23.2
matplotlib==3.1.2
tensorflow==1.15.4
Pillow==7.2.0
PyYAML==5.3.1
seaborn==0.11.1
psutil
requests